requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
google-generativeai==0.8.3
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
python-dotenv==1.0.1
tiktoken==0.8.0